file接收，Log，Ticket，没写station

+++ src/main/java/com/example/mems/entity/Ticket.java
@@ -21,10 +21,11 @@ import lombok.Setter;
 
     private static final long serialVersionUID = 1L;
 
+
       @TableId(value = "id", type = IdType.AUTO)
       private Integer id;
 
-    private Float price;
+    private  Integer price;
 
     private LocalDateTime timestamp;
